Extract node functions to separate files.
