+++ example-trace-messages/cons/query.py
@@ -1,0 +1,3 @@
+def query(query_str: str) -> str:
+    """Perform the actual query (placeholder)."""
+    return f"Response to: {query_str}"
