+++ example-trace-messages/cons/response_to_markdown.py
@@ -1,0 +1,3 @@
+def response_to_markdown(response: str) -> str:
+    """Convert the response to markdown format."""
+    return f"# Response\n\n{response}"
