+++ example-trace-messages/cons/stdout.py
@@ -1,0 +1,4 @@
+def stdout(value: str) -> str:
+    """Print the value to stdout and return it unchanged."""
+    print(value)
+    return value
